Pull out more App.scala
